+++ docs/Changelog.md
@@ -43,6 +43,7 @@ Nussknacker versions
 * [#1543](https://github.com/TouK/nussknacker/pull/1543) `ComponentProvider` API enables adding new extensions without changing e.g. `ProcessConfigCreator`
 * [#1471](https://github.com/TouK/nussknacker/pull/1471) Initial version of session window aggregate added (API may change in the future).
 * [#1631](https://github.com/TouK/nussknacker/pull/1631) Ability to use multiple config files with `nussknacker.config.location` system property
+* [#1512](https://github.com/TouK/nussknacker/pull/1512) Add kafka source `KafkaGenericNodeSourceFactory` with additional variable with metadata.
 
 0.3.1 (not released yet) 
 ------------------------